~/A2A_Demos/requirements.txt
a2a-sdk[http-server]>=0.3.25
httpx>=0.27.0
uvicorn>=0.30.0
openai>=1.0.0
fastmcp>=2.0.0
